+++ plone/app/event/ical/exporter.py
@@ -320,10 +320,13 @@ class EventsICal(BrowserView):
         return cal.to_ical()
 
     def __call__(self):
-        name = '%s.ics' % self.context.getId()
-        self.request.RESPONSE.setHeader('Content-Type', 'text/calendar')
-        self.request.RESPONSE.setHeader(
+        ical = self.get_ical_string()
+        name = '{0}.ics'.format(self.context.getId())
+        self.request.response.setHeader('Content-Type', 'text/calendar')
+        self.request.response.setHeader(
             'Content-Disposition',
             'attachment; filename="%s"' % name
         )
-        self.request.RESPONSE.write(self.get_ical_string())
+        self.request.response.setHeader('Pragma', 'no-cache')
+        self.request.response.setHeader('Content-Length', len(ical))
+        self.request.response.write(ical)